:recycle: Extracted theme detection and color set logic
